Adjust environment variables and move stream code
