+++ lib/validationstream.js
@@ -4,7 +4,7 @@ var mapnik = require('mapnik');
 var stream = require('stream');
 var uploadLimits = require('mapbox-upload-limits');
 var zlib = require('zlib');
-var invalid = require('../invalid');
+var invalid = require('./invalid');
 
 module.exports = ValidationStream;
 module.exports.validType = validType;
