+++ lib/validators/serialtiles.js
@@ -4,12 +4,12 @@ var fs = require('fs');
 var zlib = require('zlib');
 var invalid = require('../invalid');
 var prettyBytes = require('pretty-bytes');
-var ValidationStream = require('./validationstream');
+var ValidationStream = require('../validationstream');
 
 module.exports = validateSerialtiles;
 
 function validateSerialtiles(opts, callback) {
-  if (process.env.SkipSerialtilesValidation) return callback();
+  if (process.env.SkipVectorTileValidation) return callback();
   var limits = opts.limits || uploadLimits.serialtiles;
 
   var validationStream = ValidationStream({
